Move tokeniser out of db
