+++ usecases/schema/property_test.go
@@ -231,7 +231,7 @@ func TestHandler_AddProperty_Tokenization(t *testing.T) {
 		testCases := []testCase{}
 		for _, dataType := range dataTypes {
 			// all tokenizations
-			for _, tokenization := range helpers.Tokenizations {
+			for _, tokenization := range tokenizer.Tokenizations {
 				testCases = append(testCases, testCase{
 					dataType:             dataType,
 					tokenization:         tokenization,
@@ -266,7 +266,7 @@ func TestHandler_AddProperty_Tokenization(t *testing.T) {
 		testCases := []testCase{}
 		for _, dataType := range dataTypes {
 			// all tokenizations
-			for _, tokenization := range helpers.Tokenizations {
+			for _, tokenization := range tokenizer.Tokenizations {
 				switch tokenization {
 				case models.PropertyTokenizationWord:
 					testCases = append(testCases, testCase{
@@ -328,7 +328,7 @@ func TestHandler_AddProperty_Tokenization(t *testing.T) {
 		testCases := []testCase{}
 		for _, dataType := range dataTypes {
 			// all tokenizations
-			for _, tokenization := range helpers.Tokenizations {
+			for _, tokenization := range tokenizer.Tokenizations {
 				testCases = append(testCases, testCase{
 					dataType:     dataType,
 					tokenization: tokenization,
@@ -366,7 +366,7 @@ func TestHandler_AddProperty_Tokenization(t *testing.T) {
 		testCases := []testCase{}
 		for _, dataType := range dataTypes {
 			// all tokenizations
-			for _, tokenization := range helpers.Tokenizations {
+			for _, tokenization := range tokenizer.Tokenizations {
 				testCases = append(testCases, testCase{
 					dataType:     dataType,
 					tokenization: tokenization,
@@ -424,7 +424,7 @@ func TestHandler_AddProperty_Reference_Tokenization(t *testing.T) {
 	dataType := []string{refClass.Class}
 
 	// all tokenizations
-	for _, tokenization := range helpers.Tokenizations {
+	for _, tokenization := range tokenizer.Tokenizations {
 		propName := fmt.Sprintf("ref_%s", tokenization)
 		t.Run(propName, func(t *testing.T) {
 			_, _, err := handler.AddClassProperty(ctx, nil, &class, class.Class, false,
@@ -499,7 +499,7 @@ func Test_Validation_PropertyTokenization(t *testing.T) {
 		for _, dataType := range []schema.DataType{
 			schema.DataTypeText, schema.DataTypeTextArray,
 		} {
-			for _, tokenization := range helpers.Tokenizations {
+			for _, tokenization := range tokenizer.Tokenizations {
 				testCases = append(testCases, testCase{
 					name:             fmt.Sprintf("%s + '%s'", dataType, tokenization),
 					propertyDataType: newFakePrimitivePDT(dataType),
@@ -535,7 +535,7 @@ func Test_Validation_PropertyTokenization(t *testing.T) {
 					expectedErrMsg:   "",
 				})
 
-				for _, tokenization := range append(helpers.Tokenizations, "non_existing") {
+				for _, tokenization := range append(tokenizer.Tokenizations, "non_existing") {
 					testCases = append(testCases, testCase{
 						name:             fmt.Sprintf("%s + '%s'", dataType, tokenization),
 						propertyDataType: newFakePrimitivePDT(dataType),
@@ -559,7 +559,7 @@ func Test_Validation_PropertyTokenization(t *testing.T) {
 				expectedErrMsg:   "",
 			})
 
-			for _, tokenization := range append(helpers.Tokenizations, "non_existent") {
+			for _, tokenization := range append(tokenizer.Tokenizations, "non_existent") {
 				testCases = append(testCases, testCase{
 					name:             fmt.Sprintf("%s + '%s'", dataType, tokenization),
 					propertyDataType: newFakeNestedPDT(dataType),
@@ -582,7 +582,7 @@ func Test_Validation_PropertyTokenization(t *testing.T) {
 			expectedErrMsg:   "",
 		})
 
-		for _, tokenization := range append(helpers.Tokenizations, "non_existing") {
+		for _, tokenization := range append(tokenizer.Tokenizations, "non_existing") {
 			testCases = append(testCases, testCase{
 				name:             fmt.Sprintf("ref + '%s'", tokenization),
 				propertyDataType: newFakePrimitivePDT(""),
@@ -599,7 +599,7 @@ func Test_Validation_PropertyTokenization(t *testing.T) {
 		for _, dataType := range []schema.DataType{
 			schema.DataTypeString, schema.DataTypeStringArray,
 		} {
-			for _, tokenization := range append(helpers.Tokenizations, "non_existing") {
+			for _, tokenization := range append(tokenizer.Tokenizations, "non_existing") {
 				switch tokenization {
 				case models.PropertyTokenizationWord, models.PropertyTokenizationField:
 					testCases = append(testCases, testCase{
